Begin addressing issue #145 by moving files in the airline-web original project to the "airlineweb" package so that we can have unique package names in the original projects that will not confuse the aggregate javadoc build.
