+++ projects-parent/originals-parent/airline-web/src/main/java/edu/pdx/cs410J/airlineweb/AirlineRestClient.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original;
+package edu.pdx.cs410J.airlineweb;
 
 import com.google.common.annotations.VisibleForTesting;
 import edu.pdx.cs410J.web.HttpRequestHelper;
